+++ mkchromecast/preferences.py
@@ -66,7 +66,7 @@ def ConfigSectionMap(section):
     return dict1
 
 if tray == True:
-    from PyQt5.QtWidgets import QWidget, QLabel, QComboBox, QApplication, QPushButton
+    from PyQt5.QtWidgets import QWidget, QLabel, QComboBox, QApplication, QPushButton, QLineEdit
     from PyQt5 import QtCore
 
     class preferences(QWidget):
@@ -94,6 +94,8 @@ if tray == True:
             self.iconcolors()
             self.notifications()
             self.searchatlaunch()
+            if platform == 'Linux':
+                self.alsadevice()
             self.buttons()
             self.window()
 
@@ -105,16 +107,19 @@ if tray == True:
                 'node',
                 'ffmpeg',
                 'avconv',
-                'parec'
+                'parec',
+                'gstreamer'
                 ]
             self.backends = []
             if platform == 'Darwin':
                 for item in backends_supported:
-                    if is_installed(item) == True and item != 'avconv':
+                    if is_installed(item) == True and item != 'avconv' and item !='gstreamer':
                         self.backends.append(item)
             else:
                 for item in backends_supported:
-                    if is_installed(item) == True and item != 'node':
+                    if is_installed(item) == True and item != 'node' and item != 'gstreamer':
+                        self.backends.append(item)
+                    elif is_installed('gst-launch-1.0') == True and item == 'gstreamer': # Harcoded gst-launch-1.0 for gstreamer
                         self.backends.append(item)
             backendindex = self.backends.index(self.backendconf)
             self.backend = QLabel('Select Backend', self)
@@ -192,6 +197,7 @@ if tray == True:
             Sample rate
             """
             self.samplerates = [
+                '192000',
                 '96000',
                 '48000',
                 '44100',
@@ -267,20 +273,34 @@ if tray == True:
             self.qcatlaunch.setCurrentIndex(launchindex)
             self.qcatlaunch.activated[str].connect(self.onActivatedatlaunch)
 
+        def alsadevice(self):
+            """
+            Set the ALSA Device
+            """
+            self.alsadevice = QLabel('ALSA Device', self)
+            self.alsadevice.move(20*self.scale_factor, 244*self.scale_factor)
+            self.qle = QLineEdit(self)
+            self.qle.move(179*self.scale_factor, 244*self.scale_factor)
+            self.qle.setFixedWidth(84*self.scale_factor)
+            self.read_defaults()
+            if self.alsadeviceconf != None:
+                self.qle.setText(self.alsadeviceconf)
+            self.qle.textChanged[str].connect(self.onActivatedalsadevice)
+
         def buttons(self):
             """
             Buttons
             """
             resetbtn = QPushButton("Reset Settings", self)
-            resetbtn.move(10*self.scale_factor, 252*self.scale_factor)
+            resetbtn.move(10*self.scale_factor, 274*self.scale_factor)
             resetbtn.clicked.connect(self.reset_configuration)
 
             faqbtn = QPushButton("FAQ", self)
-            faqbtn.move(138*self.scale_factor, 252*self.scale_factor)
+            faqbtn.move(138*self.scale_factor, 274*self.scale_factor)
             faqbtn.clicked.connect(lambda: webbrowser.open('https://github.com/muammar/mkchromecast/wiki/FAQ'))
 
             donbtn = QPushButton("Donate :)", self)
-            donbtn.move(204*self.scale_factor, 252*self.scale_factor)
+            donbtn.move(204*self.scale_factor, 274*self.scale_factor)
             donbtn.clicked.connect(lambda: webbrowser.open('https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=JQGD4UXPBS96U'))
 
         def window(self):
@@ -289,9 +309,9 @@ if tray == True:
             """
             self.setGeometry(300*self.scale_factor, 300*self.scale_factor, 300*self.scale_factor, 200*self.scale_factor)
             if platform == 'Darwin':
-                self.setFixedSize(310*self.scale_factor, 300*self.scale_factor)     #This is to fix the size of the window
+                self.setFixedSize(310*self.scale_factor, 320*self.scale_factor)     #This is to fix the size of the window
             else:
-                self.setFixedSize(282*self.scale_factor, 300*self.scale_factor)     #This is to fix the size of the window
+                self.setFixedSize(282*self.scale_factor, 320*self.scale_factor)     #This is to fix the size of the window
             self.setWindowFlags(QtCore.Qt.WindowCloseButtonHint | QtCore.Qt.WindowMinimizeButtonHint | QtCore.Qt.WindowStaysOnTopHint)
             self.setWindowTitle('Mkchromecast Preferences')
 
@@ -336,14 +356,14 @@ if tray == True:
 
         def onActivatedbk(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','backend',text)
+            self.config.set('settings', 'backend', text)
             self.write_config()
             self.read_defaults()
             self.qccodec.clear()
             if self.backendconf == 'node':
                 codecs = ['mp3']
                 self.config.read(self.configf)
-                self.config.set('settings','codec','mp3')
+                self.config.set('settings', 'codec', 'mp3')
                 self.write_config()
             else:
                 codecs = [
@@ -365,7 +385,7 @@ if tray == True:
 
         def onActivatedcc(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','codec',text)
+            self.config.set('settings', 'codec', text)
             self.write_config()
             self.read_defaults()
             self.qcbitrate.clear()
@@ -396,31 +416,40 @@ if tray == True:
 
         def onActivatedbt(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','bitrate',text)
+            self.config.set('settings', 'bitrate', text)
             self.write_config()
             self.read_defaults()
 
         def onActivatedsr(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','samplerate',text)
+            self.config.set('settings', 'samplerate', text)
             self.write_config()
             self.read_defaults()
 
         def onActivatednotify(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','notifications',text)
+            self.config.set('settings', 'notifications', text)
             self.write_config()
             self.read_defaults()
 
         def onActivatedcolors(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','colors',text)
+            self.config.set('settings', 'colors', text)
             self.write_config()
             self.read_defaults()
 
         def onActivatedatlaunch(self, text):
             self.config.read(self.configf)
-            self.config.set('settings','searchatlaunch',text)
+            self.config.set('settings', 'searchatlaunch', text)
+            self.write_config()
+            self.read_defaults()
+
+        def onActivatedalsadevice(self, text):
+            self.config.read(self.configf)
+            if not text:
+                self.config.set('settings', 'alsadevice', None)
+            else:
+                self.config.set('settings', 'alsadevice', text)
             self.write_config()
             self.read_defaults()
 
@@ -429,24 +458,25 @@ if tray == True:
             self.codecconf = ConfigSectionMap('settings')['codec']
             if self.backendconf == 'node' and self.codecconf != 'mp3':
                 self.config.read(self.configf)
-                self.config.set('settings','codec','mp3')
+                self.config.set('settings', 'codec', 'mp3')
                 self.write_config()
                 self.codecconf = ConfigSectionMap('settings')['codec']
             self.bitrateconf = ConfigSectionMap('settings')['bitrate']
             self.samplerateconf = ConfigSectionMap('settings')['samplerate']
             self.notificationsconf = ConfigSectionMap('settings')['notifications']
-            self.searchatlaunchconf = ConfigSectionMap('settings')['searchatlaunch']
             self.searchcolorsconf = ConfigSectionMap('settings')['colors']
+            self.searchatlaunchconf = ConfigSectionMap('settings')['searchatlaunch']
+            self.alsadeviceconf = ConfigSectionMap('settings')['alsadevice']
             if debug == True:
                 print(self.backendconf, self.codecconf, self.bitrateconf, \
                         self.samplerateconf, self.notificationsconf, \
-                        self.searchatlaunchconf, self.searchcolorsconf)
+                        self.searchatlaunchconf, self.searchcolorsconf, \
+                        self.alsadeviceconf)
 
         def write_config(self):
             """This method writes to configfile"""
             with open(self.configf, 'w') as configfile:
                     self.config.write(configfile)
-
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)